fixed mocks so 'make test' works
